Perform didInsertElement behaviors afterRender
This works around a major issue in my app where navigating into a route
containing the `{{pl-uploader}}` component casuses a rerender loop which
ultimatly crashes the browser.

Pretty much did the minimum here to make my app work. Not sure the best
approach for long term glimmer support.

@rwjblue took a look at this behavior and was going to try and isolate
it so it can be fixed upstream in Ember.