Refresh the uploader on queue change
This solves the issue where the browse button moves
around, like if it's at the bottom of the page, and you display
the added files, and ass you add more files the button moves down.
Also solves the problem where you remove all files.

This keeps the browse button and any UI in the dropzone usable.