use a global drag listener for handling drag and drop
This normalizes dragenter / dragleave events so they work
as expected. Hopefully, this removes problems where there's
no matching dragleave for a dragenter, and deduplicates any
events that may happen.

Dropping on a page will now result in a no-op, which means a user
won't lose any work if they miss a dropzone.

In addition, this should reduce memory usage for apps
that have lots of dropzones.