🐛 fix polyfill uuid()
Previously, the polyfill version of `uuid()` would return strings of the form
`04aNaNaN-aNaN-40aN-80aN-aNaNaNaNaNaN`. That's because the polyfill `random()`
was returning very sparse arrays because of an early `return`. Moving the
`return` to after the `for` loop completes fully populates the arrays, resulting
in `uuid`s of the form `335dcdbe-a5e3-4409-9ecf-439d4069ea35`.

The test suite was intermittently failing because the very low entropy in
`04aNaNaN-aNaN-40aN-80aN-aNaNaNaNaNaN` meant that

```js
assert.notEqual(uuid.short(), uuid.short());
```

would sometimes fail.