Fix MIME type checks (#231)
Using a regular expression to check MIME types was problematic and unnecessary. Problematic because it wasn't escaping characters, so for example, `image/svg+xml` would not match because the `+` was interpreted as a 1-or-more modifier on the `g`. Unnecessary because the accept attribute only accepts a very limited set of clearly enumerated type wildcards (see https://html.spec.whatwg.org/multipage/input.html#file-upload-state-(type=file)).

So, match more tightly and correctly doing string comparisons, and add some unit tests for this (and to verify case-insensitivity, which was missing for mime types).