Clear mirage deprecations
In v3.x functions from miragejs will no longer be exported by mirage and
must be pulled directly from miragejs.

The configuration system has been modified to match the upstream
expectations and by more explicit as well.

Using embroider macros we can detect if ember-cli-mirage and miragejs
are present and only include this code when that is true. For our own
local tests we then add mirage as a dev dependency and require it as an
optional peer dependency so embroider will know how it should work.