Rename FileReader to UploadFileReader and convert to typescript (#759)
Similar to what we found with the File class. FileReader is an existing browser API and having this utility named the same is misleading.

Since v5 this utility is no longer documented public API. I also don't see a reason for addon consumers to use it as there are new async Blob APIs shipping with browsers.

It's used internally and in the mirage handler. In future we could entirely remove this.