Refactor queued files autotracking test (#782)
Test temporarily disabled in 519f6a9 as there's a hard to isolate timing change in a dependency which was making it fail in some matrix builds.

Refactored to mock the state change of the file via a simulated upload rather than using the mirage handler. This isolates the part we actually want to test here – tracking of files in queues.