Listen for drag-and-drop events on the Dropzone element rather than document (#793)
Think I was able to reproduce the reported issue using a regression test.

Refactored DragListener to attach and remove all events from the Dropzone element itself.

In the pre-Octane design there was a global DragListener. This would be more efficient for a page with many Dropzones and could be addressed in a further refactor. For now though, this does not change performance, and hopefully fixes the reported issue (Fixes #446).

Also replaced some this context binding with the action decorator as I'm not sure the event listener removal was using the same functions as the event listener addition.