Mirage hander: Don't send error response when file metadata can't be generated (#797)
Instead the boolean property hasAdditionalMetadata reports whether additional metadata was extracted.

After checking #152 I agree with the issue reporter. Throwing an error response from our provided handler because metadata cannot be generated by the brower's environment is unhelpful to users.

Users of the handler are free to respond with any response codes and errors themselves via the handler's own API.

With the new hasAdditionalMetadata prop it would be trivial to reimplement the "respond with 500 if metadata extraction fails" functionality in user land.

Includes dumped files and metadata matchers to ensure all metadata pipelines are working.