Migrate addon to v2 (#800)
Mirage handler

I have a ready to go branch downstream of this which extracts the mirage handler into a new ember-file-upload-mirage package.

Moving the mirage handlers to their own package will be a breaking change and will be included with this change as a major release.

Test waiter
I think that because we import UploadFile within the test-app the test waiter declaration is evaluated twice – results in a console warning. To at least avoid this I've replaced buildWaiter with waitForPromise which is run inline rather than in module scope anyway.


TypeScript interfaces

I've moved most ts interfaces into a dedicated interfaces file. This is because importing across different modules creates a bit of a wild import map and can result in circular dependencies. I was advised to try this on Discord and it's working well.

It's (almost) compiled away by rollup, but it does contain some enums which are imported in the test-app so it's currently a public export. We could fix this with my next suggestion anyway...


Module map

Since we're releasing breaking changes here – this could be a chance to improve the ergonomics of module exports.

I'd be in favour of named exporting all non-ember modules from ember-file-upload. To be included as a breaking change in v7. Can do this in a follow-up PR.