Add @ember/string as a dependency of ember packages (#916)
It's a required peer dependency since ember-source 4.10.0. The deprecation is now a removal in canary v5 which is causing a CI failure.

Same change would have been made in the ember-cil 4.12 blueprint but hoping to get the CI green with a minimal change first.

Related to #899 as I'm not sure we need to declare a peerDependency from the addon itself. Reason for this is don't think the addon uses @ember/string.