Bump build dependencies (#920)
Updates to typescript, rollup, rollup-plugin-ts.

There seems to be a combination of these packages that is failing the rollup build (in the lockfile update PR, for example).

Pushing these packages up (dev dependencies only) results in working builds and no typescript errors.

Also all of the dependency update PRs are failing due to apparently bad lockfiles. Once this is merged I'll rebase them.