Add `onUploadStarted`, `onUploadSucceeded` and `onUploadFailed` to `file-queue` helper  (#925)
Re-staged version of #910 since that PR was getting messy. Should be easier to review this one.

Adds callbacks to file-queue helper:

onUploadStarted
onUploadSucceeded
onUploadFailed

So that UploadFile instances have a sensible state at callback time I have adjusted the queue flush from a side-effect of the File.state setter and instead call it explicitly. As I understand this is an internal change, only.

Should fix #58