Migrate `test-app` to TypeScript. Add Typecheck job to CI (#944)
Main purpose is to add coverage for TypeScript types which we are shipping but not really testing their consumption/use.

Secondary purpose is to prepare infrastructure for testing and shipping Glint types.

Some internal system tests remain untyped as they will be more difficult to type. I'm ok with this for now as they aren't public API. It may be easier to refactor than them.