Relax peer dependency version for @ember/test-helpers (#950)
Allow @ember/test-helpers v3 to be used with this addon. v3 was released a few weeks ago and doesn’t seem to have any changes impacting this addon (release notes). I’ve been using it with ember-file-upload for a while now and haven’t had any issues, so recommend relaxing the peer requirements so npm won’t protest when trying to install the latest test-helpers version.