Remove reliance on ember-cli-mirage (#1102)
ember-cli-mirage is an optional dependency and not actually required for the uploadHandler to function. As shown in the docs, the user will need to set up the Mirage server themselves for this to work at all but that server can be achieved through vanilla miragejs, as shown in this issue: bgantzler/ember-mirage#60

This PR removes the requirement on ember-cli-mirage.